Used struct to define characters. Players character list view is now more separated from the model, not totally
