+++ OC_FrenchGameFactory/OC_FrenchGameFactory/Classes/Character.swift
@@ -7,16 +7,10 @@
 
 import Foundation
 
-class Character{
+struct Character{
     var originalName : String
     var customName: String
     var health: Int
     var weapon: Weapon
     var chestAvailable: Bool?
-    init(originalName : String, customName: String, health: Int, weapon: Weapon){
-        self.originalName = originalName
-        self.health = health
-        self.weapon = weapon
-        self.customName = customName
-    }
 }
