+++ OC_FrenchGameFactory/OC_FrenchGameFactory/Controller/ViewController.swift
@@ -8,15 +8,15 @@
 import UIKit
 
 class ViewController: UIViewController {
-    @IBOutlet weak var player1CharacterList: UIStackView!
-    @IBOutlet weak var player2CharacterList: UIStackView!
+    @IBOutlet weak var player2CharacterList: CharacterList!
+    @IBOutlet weak var player1CharacterList: CharacterList!
     
     override func viewDidLoad() {
         super.viewDidLoad()
         
 //        Populate the empty character list for each players when the app starts
-        CharacterList.displayAvailableCharacters(in: player1CharacterList)
-        CharacterList.displayAvailableCharacters(in: player2CharacterList)
+        player1CharacterList.displayAvailableCharacters()
+        player2CharacterList.displayAvailableCharacters()
         
     }
 }
