+++ OC_FrenchGameFactory/OC_FrenchGameFactory/Model/Characters/Kiwi.swift
